Implemented user stories around taking turns and selection.
 - As a player, I can take my turn, alternatively with my opponent.
 - As a player, I am unable to move while my opponent is moving.
 - As a player, I can tell when it is my move.
 - As a player, I can make 3 moves per turn.
 - As a player, I can deselect a piece by clicking on the same 
   square again.
   
To do this, I also had to switch from using a simple string to 
represent each piece to an actual piece object. I added some basic
other fields such a class may have, but its still really only 
using who it belongs to and what type of piece it is.